Fix ability to import ripgrep into munki
At present, when the current munki recipe is run an error is thrown because munkiimport can't find a valid item in the DMG since the entire Github master folder is inside.  Even if pointed to the executable, munki doesn't know how to process it.  This addresses the problem by creating a pkg with a new pkg recipe and have the munki recipe import this instead.

- Add .pkg recipe to extract and copy the rg executable in /usrl/local/bin/ to make a pkg
- Update .munki recipe to use new parent recipe, remove no longer needed Unarchiver and DmgCreator procesors, and use pkg at %pkg_path%. Update description.
- Add MUNKI_CATEGORY variable for category pkginfo key